refactor : refactor package structure
