+++ HolubLife/src/com/holub/life/view/menu/ClockMenu.java
@@ -1,8 +1,6 @@
-package com.holub.ui;
+package com.holub.life.view.menu;
 
 import com.holub.life.controller.Clock;
-import com.holub.life.controller.Universe;
-import com.holub.life.view.UniverseView;
 
 import javax.swing.*;
 import java.awt.event.ActionEvent;
