+++ HolubLife/src/com/holub/life/view/menu/GridMenu.java
@@ -1,9 +1,8 @@
-package com.holub.ui;
+package com.holub.life.view.menu;
 
 import com.holub.life.controller.Universe;
 import com.holub.life.view.UniverseView;
 
-import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
