+++ HolubLife/src/com/holub/life/view/menu/UndoMenu.java
@@ -1,4 +1,4 @@
-package com.holub.ui;
+package com.holub.life.view.menu;
 
 import com.holub.life.controller.Universe;
